Improved readability

+++ checkout/src/main/kotlin/com/davromalc/checkout/dto/CheckoutDTO.kt
@@ -2,4 +2,4 @@ package com.davromalc.checkout.dto
 
 import java.math.BigDecimal
 
-data class CheckoutDTO (val cartId: String, val amount : BigDecimal)
+data class CheckoutDTO (val cartId: String, val amount : Double)
